Wire in moderation and bleaching to new models.
In doing so, I've replaced the explicit fields for rendered HTML with
properties that render lazily, using the cache to store results. This
should avoid the potential for inconsistent data and mean we can update
bleach settings without having to update everything in the database.

Could probably do with a couple more tests.