Use an environment variable to choose settings.
This gets around the issue when changing the import in settings_local
where tests depending on the Betafarm URL structure would fail. I'm sure
there are more elegant solutions to this problem, but I'm not sure they
fit in with the existing assumption that the local settings module
imports everything from the base settings.