Add a login page.
In the future, we are either going to integrate this with the submission
form or flesh it out with some explanation about why we're using
BrowserID and how it works. For now, this just means we can use the
login_required decorator without breaking things.

The new context processor make sure the BrowserID form always takes the
user back to the page they were on, unless it was the login form, in
which case it will follow through to the appropriate 'next' page.