Add a first test for the edit entry view.
This involved wiring it, in a basic capacity, into the Betafarm site
structure, even if it's only going to be used on Ignite, because the
tests take place in the context of Betafarm.

I also copied the URL middleware suppression code from commit
f770093b16; once these are merged, we should be able to eliminate one or
other of the declarations.