Fix up tests to accommodate links and errors.
The external links form requires a few extra bits of plumbing in the
submission data, which the tests weren't supplying. Also, the error keys
had changed and the tests weren't updated to match.

To my knowledge, all tests are now passing except for the one that is
currently being skipped.